feat(venn): SJIP-1254 rename venn chart to index
